refactor(pkg): rename eventskafka -> eventsmessaging/kafkapub
Lays the groundwork for adding a NATS JetStream sibling
(pkg/eventsmessaging/natspub). Behaviour unchanged.
